requirements.txt
pytest-playwright==0.5.1
# playwright==1.45.1
pydantic==2.8.2
fastapi==0.111.1
SQLAlchemy==2.0.31
uvicorn==0.30.3